The most critical fix in existence. (#8)
Logs both a Unity Error and a System Debug report when a configured type GUID is not found in the type registry.  This can be caused by either the requested type not being in the project (e.g. in a UPM package not registered) or, MORE LIKELY, Unity code stripping has deemed the class NOT WORTHY and stripped it out.
Comments and Logs point to creating a Link.XML to stop unity being a GIT.